Split URL Helper tests
